refactor(tui): split add flow field groups
