+++ .sampo/changesets/1055-add-flow-field-groups.md
@@ -1,0 +1,5 @@
+---
+npm/wp-typia: patch
+---
+
+Split interactive add flow field rendering into focused add-kind field groups.
